pyproject.toml
[project]
name = "mono-ai-service-starter"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriele Lanzafame",email = "gabriele.lanzafame03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.0,<0.117.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (>=1.93.1,<2.0.0)",
    "wandb (>=0.21.0,<0.22.0)",
    "torch (>=2.7.1,<3.0.0)",
    "torchaudio (>=2.7.1,<3.0.0)",
    "torchvision (>=0.22.1,<0.23.0)",
    "transformers (>=4.53.1,<5.0.0)",
    "peft (>=0.16.0,<0.17.0)",
    "bitsandbytes (>=0.46.1,<0.47.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "trl (>=0.19.0,<0.20.0)",
    "weave (>=0.51.54,<0.52.0)",
    "gql[aiohttp] (>=3.4.0,<4.0.0)",
    "gpl[aiohttp] (>=0.1.4,<0.2.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "accelerate (>=1.8.1,<2.0.0)",
]

[tool.poetry]
package-mode = false 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"

